# Prototype of codegen for ArrowWriter.